Begin building master side of app
This commit cannot render anything, it can only create/list/delete jobs from a master role. All of this data is saved in a folder called `render-farm` in the working directory in a data file. This way all the jobs can be accessed by all the clients.

I will clean up some of this code in the next commit, but I need to sleep now