Add support for waiting for work and fix previous commit
Apparently blender saves .jpg files when rendering in the "JPEG" mode, even though it refuses to take "JPG"...

Now when the app has finished going through the jobs, it will check if there are any frames that need to be rendered. If there are not, if will begin to watch each job's output directory as well as the working 'render-farm' directory. If any files are deleted or updated (i.e. a frame needs to be rerendered or a new job is added), the app will check if it needs to render again and will render if necessary.
This way you can just start the application on a computer and just have it continuously wait until it is needed without having to ever restart the application